requirements.txt
pandas 
numpy
streamlit
tensorflow-cpu
matplotlib
seaborn
Pillow
